Refactor useDefaultState into two separate hooks
