+++ src/common/components/SearchField.tsx
@@ -4,6 +4,7 @@ import InputAdornment from '@material-ui/core/InputAdornment';
 import Search from '@material-ui/icons/Search';
 
 import useDefaultState from 'common/hooks/useDefaultState';
+import useCallbackOnChange from 'common/hooks/useCallbackOnChange';
 
 interface SearchFieldProps {
   defaultValue?: string;
@@ -24,15 +25,16 @@ const SearchField = ({
   fullWidth,
   className,
 }: SearchFieldProps) => {
-  const [text, setText, handleSearch] = useDefaultState(defaultValue ?? '', onSearch);
+  const [text, setText] = useDefaultState(defaultValue ?? '');
+  const handleSearch = useCallbackOnChange(defaultValue ?? '', onSearch);
 
   const handleChange = (event: React.ChangeEvent<HTMLInputElement>): void =>
     setText(event.currentTarget.value);
 
   const handleKeyDown = (event: React.KeyboardEvent<HTMLInputElement>): void =>
-    event.key === 'Enter' ? handleSearch() : undefined;
+    event.key === 'Enter' ? handleSearch(text) : undefined;
 
-  const handleBlur = (): void => (searchOnBlur ? handleSearch() : undefined);
+  const handleBlur = (): void => (searchOnBlur ? handleSearch(text) : undefined);
 
   return (
     <TextField
