+++ src/common/hooks/useDefaultState.ts
@@ -1,34 +1,17 @@
-import { useState, useEffect, useRef } from 'react';
-import _ from 'lodash';
+import { useState, useEffect } from 'react';
 
 /**
- * Returns a stateful value, a function to update it, and a function which invokes the given
- * callback only when the stateful value has changed.
- *
- * @param defaultValue Dynamically changing default value.
- * @param callback Invoked when the returned handler function is called and the stateful value has since changed.
+ * Returns a stateful value and a function to update it. Unlike useState, the default value can
+ * change dynamically.
  */
-const useDefaultState = <T>(
-  defaultValue: T,
-  callback: (value: T) => void
-): [T, (arg: T) => void, () => void] => {
+const useDefaultState = <T>(defaultValue: T): [T, (arg: T) => void] => {
   const [value, setValue] = useState(defaultValue);
-  const prevValue = useRef(defaultValue);
 
   useEffect(() => {
-    prevValue.current = defaultValue;
     setValue(defaultValue);
   }, [defaultValue]);
 
-  const handler = (): void => {
-    // Only invoke callback if value has changed since last callback invocation.
-    if (!_.isEqual(value, prevValue.current)) {
-      prevValue.current = value;
-      callback(value);
-    }
-  };
-
-  return [value, setValue, handler];
+  return [value, setValue];
 };
 
 export default useDefaultState;
